+++ components/Column/index.tsx
@@ -42,6 +42,7 @@ export type ColumnStyleProps = {
   shrink?: boolean | number
   flex?: boolean | number
   borderWidth?: number
+  borderPosition?: 'top' | 'right' | 'bottom' | 'left' | 'all'
 }
 
 export type ColumnProps<T = {}> = T &
@@ -80,6 +81,7 @@ export const withColumnStyleProps = <T,>(
       shrink,
       flex,
       borderWidth,
+      borderPosition = 'all',
       ...nonStyleProps
     } = props
 
@@ -93,7 +95,6 @@ export const withColumnStyleProps = <T,>(
       }),
       ...(textColor && { color: getColorShade(textColor) }),
       ...(borderColor && { borderColor: getColorShade(borderColor) }),
-      ...(borderWidth !== undefined && { borderWidth }),
       ...(unitsAround !== undefined && { margin: getUnits(unitsAround) }),
       ...(unitsAbove !== undefined && { marginTop: getUnits(unitsAbove) }),
       ...(unitsBelow !== undefined && { marginBottom: getUnits(unitsBelow) }),
@@ -124,6 +125,17 @@ export const withColumnStyleProps = <T,>(
       ...(alignLeft && { alignItems: 'flex-start' }),
       ...(alignCenter && { alignItems: 'center' }),
       ...(alignRight && { alignItems: 'flex-end' }),
+
+      ...(borderPosition === 'top' &&
+        borderWidth !== undefined && { borderTopWidth: borderWidth }),
+      ...(borderPosition === 'right' &&
+        borderWidth !== undefined && { borderRightWidth: borderWidth }),
+      ...(borderPosition === 'bottom' &&
+        borderWidth !== undefined && { borderBottomWidth: borderWidth }),
+      ...(borderPosition === 'left' &&
+        borderWidth !== undefined && { borderLeftWidth: borderWidth }),
+      ...(borderPosition === 'all' &&
+        borderWidth !== undefined && { borderWidth }),
     }
 
     boxShallot = applyStyles(boxShallot, shallot)